+++ src/config/field-overrides/dataApprovalWorkflow.js
@@ -4,4 +4,11 @@ export default new Map([
     ['periodType', {
         component: PeriodTypeDropDown,
     }],
+    ['categoryCombo', {
+        referenceType: 'categoryCombo',
+        fieldOptions: {
+            queryParamFilter: ['dataDimensionType:eq:ATTRIBUTE', 'name:eq:default'],
+            defaultToDefaultValue: true,
+        },
+    }],
 ]);